Improvements to JitCache (assume const) (#6)
This PR adds the `assume_const` feature to the JitCache. 

```
@triton_dejavu.jitcache(
    check_keys=["USE_ALIBI_SLOPES", "SLIDING_WINDOW", "filter_by_query_len"],
    assume_const=[
        "scale",
        "k_scale",
        "v_scale",
        "query_stride_1",
        "output_stride_1",
        "stride_k_cache_0",
        "stride_k_cache_1",
        "stride_k_cache_2",
        "stride_k_cache_4",
        "stride_v_cache_0",
        "stride_v_cache_1",
        "stride_v_cache_2",
        "stride_v_cache_2",
    ],
)
@triton.jit
def kernel_paged_attention_2d(
```

This is necessary in situations where some parameters are constant from the application PoV, but can't be declared `tl.constexpr`. This is for example the case with 64-bit addresses, which have to use `tl.int64`. 

Besides, it fixes some typos and extends the "documentation" in the Readme. 
---------

Signed-off-by: Burkhard Ringlein <ngl@zurich.ibm.com>